backend/requirements.txt
fastapi>=0.68.0
pydantic>=1.8.0
pandas>=2.2.0
pymongo>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
slowapi>=0.1.4
uvicorn>=0.15.0
catboost>=1.2.0
openai
googletrans==4.0.0-rc1
deep-translator
google-generativeai
# Windows-specific and cross-platform servers
waitress>=2.0.0
# Windows tools (optional, install as needed)
# pywin32>=300; sys_platform == 'win32'